# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Campus Connect API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/me:
    get:
      operationId: getMe
      tags: [users]
      summary: Get my profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/me/profile:
    patch:
      operationId: updateProfile
      tags: [users]
      summary: Update my profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
  /users/{userId}:
    get:
      operationId: getUserById
      tags: [users]
      summary: Get a user by ID
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User profile (privacy filtered)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicUserProfile"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/stats:
    get:
      operationId: getUserStats
      tags: [users]
      summary: Get current user stats (posts, matches, likes received)
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserStats"

  # --- COLLEGES ---
...
  /report:
    post:
      operationId: submitReport
      tags: [safety]
      summary: Report a user, post, or message
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReportRequest"
      responses:
        "201":
          description: Report submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /block:
    post:
      operationId: blockUser
      tags: [safety]
      summary: Block a user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/BlockRequest"
      responses:
        "200":
          description: User blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /block/{userId}:
    delete:
      operationId: unblockUser
      tags: [safety]
      summary: Unblock a user
      security:
        - bearerAuth: []
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /block/list:
    get:
      operationId: getBlockedUsers
      tags: [safety]
      summary: Get list of blocked users
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Blocked users
          content:
            application/json:
              schema:
                type: object
                properties:
                  blocked:
                    type: array
                    items:
                      $ref: "#/components/schemas/PublicUserProfile"
                required:
                  - blocked

  # --- ADMIN ---
  /admin/reports:
    get:
      operationId: getAdminReports
      tags: [admin]
      summary: Get all reports (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ReportStatus"
      responses:
        "200":
          description: Reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  reports:
                    type: array
                    items:
                      $ref: "#/components/schemas/Report"
                required:
                  - reports
  /admin/action:
    post:
      operationId: adminAction
      tags: [admin]
      summary: Take moderation action (admin only)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminActionRequest"
      responses:
        "200":
          description: Action taken
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users:
    get:
      operationId: getAdminUsers
      tags: [admin]
      summary: Get all users with flags (admin only)
      security:
        - bearerAuth: []
      parameters:
        - name: flagged
          in: query
          schema:
            type: boolean
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: object
                properties:
                  users:
                    type: array
                    items:
                      $ref: "#/components/schemas/AdminUserView"
                required:
                  - users
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
      required:
        - success
        - message

    # --- AUTH SCHEMAS ---
...
    Gender:
      type: string
      enum: [MALE, FEMALE, NON_BINARY, PREFER_NOT_TO_SAY]
    VerificationStatus:
      type: string
      enum: [PENDING, APPROVED, REJECTED]
    UserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
          nullable: true
        email:
          type: string
        bio:
          type: string
          nullable: true
        gender:
          $ref: "#/components/schemas/Gender"
          nullable: true
        age:
          type: integer
          nullable: true
        profilePhotos:
          type: array
          items:
            type: string
        collegeId:
          type: string
        college:
          $ref: "#/components/schemas/College"
          nullable: true
        interests:
          type: array
          items:
            type: string
        clubs:
          type: array
          items:
            type: string
        isEmailVerified:
          type: boolean
        verificationStatus:
          $ref: "#/components/schemas/VerificationStatus"
        isProfilePublic:
          type: boolean
        showAge:
          type: boolean
        showGender:
          type: boolean
        profileCompleted:
          type: boolean
        isActive:
          type: boolean
        lastActiveAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - profilePhotos
        - collegeId
        - interests
        - clubs
        - isEmailVerified
        - verificationStatus
        - isProfilePublic
        - showAge
        - showGender
        - profileCompleted
        - isActive
        - createdAt
    PublicUserProfile:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        username:
          type: string
          nullable: true
        bio:
          type: string
          nullable: true
        gender:
          $ref: "#/components/schemas/Gender"
          nullable: true
        age:
          type: integer
          nullable: true
        profilePhotos:
          type: array
          items:
            type: string
        collegeId:
          type: string
        college:
          $ref: "#/components/schemas/College"
          nullable: true
        interests:
          type: array
          items:
            type: string
        clubs:
          type: array
          items:
            type: string
        isEmailVerified:
          type: boolean
        profileCompleted:
          type: boolean
      required:
        - id
        - name
        - profilePhotos
        - collegeId
        - interests
        - clubs
        - isEmailVerified
        - profileCompleted
    UpdateProfileRequest:
      type: object
      properties:
        name:
          type: string
        username:
          type: string
        bio:
          type: string
        gender:
          $ref: "#/components/schemas/Gender"
        age:
          type: integer
        interests:
          type: array
          items:
            type: string
        clubs:
          type: array
          items:
            type: string
        profilePhotos:
          type: array
          items:
            type: string
        isProfilePublic:
          type: boolean
        showAge:
          type: boolean
        showGender:
          type: boolean
    UserStats:
      type: object
      properties:
        postsCount:
          type: integer
        matchesCount:
          type: integer
        likesReceived:
          type: integer
        commentsReceived:
          type: integer
      required:
        - postsCount
        - matchesCount
        - likesReceived
        - commentsReceived
    College:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        domain:
          type: string
        location:
          type: string
      required:
        - id
        - name
        - domain
        - location

    # --- POST SCHEMAS ---
...
    ReportReason:
      type: string
      enum: [SPAM, ABUSE, HARASSMENT, FAKE_PROFILE, NSFW, OTHER]
    ReportStatus:
      type: string
      enum: [PENDING, REVIEWED, ACTION_TAKEN]
    ReportRequest:
      type: object
      properties:
        targetUserId:
          type: string
        targetPostId:
          type: string
        targetMessageId:
          type: string
        reason:
          $ref: "#/components/schemas/ReportReason"
        description:
          type: string
      required:
        - reason
    BlockRequest:
      type: object
      properties:
        userId:
          type: string
      required:
        - userId
    Report:
      type: object
      properties:
        id:
          type: string
        reporterId:
          type: string
        targetUserId:
          type: string
          nullable: true
        targetPostId:
          type: string
          nullable: true
        reason:
          $ref: "#/components/schemas/ReportReason"
        description:
          type: string
          nullable: true
        status:
          $ref: "#/components/schemas/ReportStatus"
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - reporterId
        - reason
        - status
        - createdAt

    # --- ADMIN SCHEMAS ---
    AdminActionType:
      type: string
      enum: [BAN_USER, DELETE_POST, WARN_USER, DISMISS]
    AdminActionRequest:
      type: object
      properties:
        reportId:
          type: string
        actionType:
          $ref: "#/components/schemas/AdminActionType"
        notes:
          type: string
      required:
        - reportId
        - actionType
    AdminUserView:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        collegeId:
          type: string
        college:
          $ref: "#/components/schemas/College"
          nullable: true
        isActive:
          type: boolean
        isBanned:
          type: boolean
        reportCount:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - email
        - collegeId
        - isActive
        - isBanned
        - reportCount
        - createdAt