# lib/api-zod/src/generated/types/adminActionRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Campus Connect API specification
 * OpenAPI spec version: 0.1.0
 */
import type { AdminActionType } from "./adminActionType";

export interface AdminActionRequest {
  reportId: string;
  actionType: AdminActionType;
  notes?: string;
}